requirements.txt
aiogram
fastapi
aiohttp
uvicorn
pypdf2
python-docx
pydantic-settings